+++ src/routing/index.ts
@@ -2,4 +2,4 @@ export type { RouteSource } from '@/routing/registration/source/route-source';
 export { Route } from './route/route';
 export { createPathFor } from './path/create-path-for';
 export { RouteGroup } from './group/route-group';
-export { Any, Delete, Get, Head, Options, Patch, Post, Put } from './verb/http-verb-helpers';
+export { Any, Delete, Get, Head, Options, Patch, Post, Put } from '@/routing/registration/verb/http-verb-helpers';
