+++ src/routing/registration/source/normalize-route-sources.test.ts
@@ -1,5 +1,5 @@
 import { RouteGroup } from '@/routing/group/route-group';
-import { Get } from '@/routing/verb/http-verb-helpers';
+import { Get } from '@/routing/registration/verb/http-verb-helpers';
 import { describe, expect, test, vi } from 'vitest';
 import { normalizeRouteSources } from './normalize-route-sources';
