+++ src/routing/route/create-route-definition.ts
@@ -1,7 +1,7 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
+import type { HttpMethod } from '@/routing/registration/verb/http-method.type';
 import type { RouteOptions } from '@/routing/route/route-options.type';
 import type { RouterMethod } from '@/routing/route/router-method.type';
-import type { HttpMethod } from '@/routing/verb/http-method.type';
 import type { Request } from 'koa';
 import type { RouteDefinition } from './route-definition.type';
