+++ src/routing/route/route-declaration.type.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
+import type { HttpMethod } from '@/routing/registration/verb/http-method.type';
 import type { RouteOptions } from '@/routing/route/route-options.type';
-import type { HttpMethod } from '@/routing/verb/http-method.type';
 import type { Request } from 'koa';
 
 export interface RouteDeclaration<TRequest extends Request = HttpRequest> {
